Wrote the addBooking and pulled the test code directly into the file.
Note that we had to change our solution from the one from lil due to the timezone difference causing the new Date(value) to set the day a day earlier.

Note that I did have to use the solution in order to get where I needed to go because I thought the validation was happening on prop set. But, since our booking objects are set during the map function, and then converted into proxy objects, the set functions never get called, so our addBooking had to call them explicitly. This was unclear in the instructions.