changed the observerpattern to be a js file.
added the WeatherChallenge folder and the challenge and tester for the challenge.

note that my solution maintains that the observers are the direct functions to be called, while the LIL instructor stores the full object and calls a function within that object. There's a lot in the update function of the Display method taht doesn't make sense, like why pass in a displayName when it has its own display name.

I also made the observers set obselete by simply storing the boundUpdate objects in the displayNames map. this makes the displayNames variable name off.

All tests pass, but I do want to update the functionality to correctly use the displayNames variable and the observers variable, while continuing to maintain that the observers hold the bound functions, and not the whole objects.